Create 93 - Worqor Zormor.md
Updated presets for Ryoqor Terteh which freezes and draws the unsafe delayed AOE.

Updated presets for Kahderyo which draws Unsafe spots for Earthen Shot,Wind shot and Towers. It also draws the unsafe spots for Earthen Shot 2 as they are yellow and particularly hard to see.

Updated presets for Gurgurlur which draws unsafe zones on the wind and directional arrow. It also draws the unsafe first tile which blows up first, which is useful for a later knockback mechanic to avoid it.