Update 100 - Tender Valley.md
Corrected the first bosses name. Anthracite was given twice.

Included a new preset for 'Yok Huy Attestant', a miniboss before the last boss. It draws the unsafe AOEs for the laser towers off the side of the arena. This uses vfx and distance limit. Example of one pattern of the draw is here. There are 3 patterns total, 12 towers all together that have been drawn.